requirements.txt
asttokens 
certifi==2025.1.31
charset-normalizer==3.4.1
comm 
contourpy==1.3.1
cycler==0.12.1
debugpy 
decorator 
exceptiongroup 
executing 
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata 
ipykernel 
ipython 
jedi 
Jinja2==3.1.5
jupyter_client 
jupyter_core 
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline 
mpmath==1.3.0
nest_asyncio 
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging 
parso 
pexpect 
pickleshare 
pillow==11.1.0
platformdirs 
prompt_toolkit 
psutil 
ptyprocess 
pure_eval 
Pygments 
pyparsing==3.2.1
python-dateutil 
PyYAML==6.0.2
pyzmq 
requests==2.32.3
safetensors==0.5.3
six 
stack_data
sympy==1.13.1
timm==0.6.12
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado 
tqdm==4.67.1
traitlets 
triton==3.2.0
typing_extensions
urllib3==2.3.0
wcwidth 
zipp 
